Fix linux build
Fix variant_error include
Add SConstruct build script
GCC template compatibility fix
Fix preprocessor error, default is int anyways
Explicit this to make gcc happy
Use uint8_t since chars signedness is undefined
Add a few typenames
Implement make_copy in a standard compliant way
Remove no longer necessary boost::filesystem::native specifier
Move boost::tuples::length out of boost::format to avoid linker errors
Remove default argument to match previous declaration and fix gcc warning
Add ifdefs to support compilation with msvc and gcc
Use new typedef for all compilers
Remove functions without implementations
Forward declare iterator types for all compilers
Remove typedefs that changed the meaning of types
Changes for VS 2015 and Boost 1.6.1 compatibility
Use difference_type specified in traits
Work around the "incomplete type" error from boost iterator_facade
Use "NULL" in string representation
Use char type in base64 conversions
Remove iterator_category from container
Set double precision in stream based on numeric_limits
Set compiler optimsation settings
Copy thirdy-party dependencies to local directory
Compare file sizes and dates in lazy-copy logic

+++ protean/detail/xml_writer_impl.hpp
@@ -44,7 +44,7 @@ namespace protean { namespace detail {
 
         struct element_info
         {
-            element_info(const std::string& name, int level) :
+            element_info(const std::string& name, size_t level) :
                 m_name(name),
                 m_attributes(variant::Dictionary),
                 m_level(level)
